Move the ContentMetadataGenerator to the Cocina::ToFedora module
This place is reflective of the classes purpose
